added dempstersrule logic
