+++ DempsterRule.py
@@ -1,12 +1,19 @@
-from BasisMeasure import BasisMeasure
+from Evidence import Evidence
 
 class DempsterRule:
-    def __init__(self, bm1, bm2):
-        self.bm1 = bm1
-        self.bm2 = bm2
-        self.bmout = BasisMeasure(None, None)
+    def __init__(self, ev1, ev2):
+        self.ev1 = ev1
+        self.ev2 = ev2
+        self.evout = Evidence(None, None)
 
-        for x in self.bm1.getentries():
-            for y in self.bm2.getentries():
+        for x in self.bm1.get_entries():
+            for y in self.bm2.get_entries():
+                if x[0] or y[0] is 'omega':
+                    prob = 
+                    # set bmout value equals the other one
+                if x and y is empty:
+                    # calc
+                else:
+                    unify and do stuff
 
         # bmout.addentry
